handlers/health_test.go
package handlers

import (
    "testing"
    "net/http"
    "net/http/httptest"
    "encoding/json"
)

func TestHealthHandler(t *testing.T) {
    // Create a request to pass to our handler. We don't have any query parameters for now, so we'll
    // pass 'nil' as the third parameter.
    req, err := http.NewRequest("GET", "/health", nil)
    if err != nil {
        t.Fatal(err)
    }

    // We create a ResponseRecorder (which satisfies http.ResponseWriter) to record the response.
    rr := httptest.NewRecorder()
    handler := http.HandlerFunc(Health)

    // Our handlers satisfy http.Handler, so we can call their ServeHTTP method
    // directly and pass in our Request and ResponseRecorder.
    handler.ServeHTTP(rr, req)

    // Check the status code is what we expect.
    if status := rr.Code; status != http.StatusOK {
        t.Errorf("handler returned wrong status code: got %v want %v",
            status, http.StatusOK)
    }

    var response = make(map[string]string)
    json.Unmarshal(rr.Body.Bytes(), &response)
    if response["alive"] != "true"{
        t.Errorf("handler returned unexpected body: got %v", response)
    }
}
